Added soundmanager
The soundmanager creates a filereader which looks for two folders: “noCopySounds” and “otherSounds” inside the folder from which the program executes. If these folders do not exist, the filereader will create them. The filereader will read the contents of both of these folders and place files that ends with “.wav” into two different string lists (the path is interesting here).

The soundmanager will check both of these lists and create SoundEffects with the help of FileStream. The instances ends up in two different lists of type SoundEffectInstance. We can then play the sounds from these directorys inside our game. (Y)